@@ -4,7 +4,7 @@ import (
 	"encoding/base32"
 	"encoding/base64"
 	"encoding/json"
-	"github.com/satori/go.uuid"
+	"github.com/google/uuid"
 	"net/http"
 )
 
@@ -15,24 +15,27 @@ var (
 	idEncoding   = base64.StdEncoding.WithPadding(base32.NoPadding)
 )
 
-func NewId(fromOtherReverseProxy bool, req *http.Request) Id {
+func NewId(fromOtherReverseProxy bool, req *http.Request) (Id, error) {
 	return newId(fromOtherReverseProxy, "X-Request-ID", req)
 }
 
-func NewCorrelationId(fromOtherReverseProxy bool, req *http.Request) Id {
+func NewCorrelationId(fromOtherReverseProxy bool, req *http.Request) (Id, error) {
 	return newId(fromOtherReverseProxy, "X-Correlation-ID", req)
 }
 
-func newId(fromOtherReverseProxy bool, fromHeader string, req *http.Request) Id {
+func newId(fromOtherReverseProxy bool, fromHeader string, req *http.Request) (Id, error) {
 	if fromOtherReverseProxy {
 		if x := req.Header.Get(fromHeader); len(x) > 0 && len(x) <= 256 {
 			if id, err := ParseId(x); err == nil {
-				return id
+				return id, nil
 			}
 		}
 	}
-	val := uuid.NewV4()
-	return Id(val)
+	val, err := uuid.NewRandom()
+	if err != nil {
+		return Id{}, err
+	}
+	return Id(val), nil
 }
 
 func (instance Id) String() string {
@@ -44,7 +47,7 @@ func (instance Id) MarshalJSON() ([]byte, error) {
 }
 
 func ParseId(plain string) (Id, error) {
-	val, err := uuid.FromString(plain)
+	val, err := uuid.Parse(plain)
 	if err != nil {
 		return Id{}, err
 	}
