@@ -3,7 +3,7 @@ package definition
 import (
 	"fmt"
 	"github.com/echocat/lingress/support"
-	log "github.com/sirupsen/logrus"
+	"github.com/echocat/slf4g"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/client-go/tools/cache"
@@ -13,6 +13,7 @@ import (
 
 type Definition struct {
 	MaxTries int
+	Logger   log.Logger
 
 	typeDescription string
 	informer        cache.SharedInformer
@@ -29,10 +30,11 @@ type OnElementUpdatedFunc func(key string, old, new metav1.Object) error
 type OnElementRemovedFunc func(key string) error
 type OnErrorFunc func(key string, event string, err error)
 
-func newDefinition(typeDescription string, informer cache.SharedInformer) (*Definition, error) {
+func newDefinition(typeDescription string, informer cache.SharedInformer, logger log.Logger) (*Definition, error) {
 	return &Definition{
 		typeDescription: typeDescription,
 		informer:        informer,
+		Logger:          logger.With("type", typeDescription),
 	}, nil
 }
 
@@ -45,11 +47,14 @@ func (instance *Definition) Init(stop support.Channel) error {
 		panic(fmt.Sprintf("definition %s has no informer", instance.typeDescription))
 	}
 
-	instance.informer.AddEventHandler(cache.ResourceEventHandlerFuncs{
+	_, err := instance.informer.AddEventHandler(cache.ResourceEventHandlerFuncs{
 		AddFunc:    instance.onClusterElementAdded,
 		UpdateFunc: instance.onClusterElementUpdated,
 		DeleteFunc: instance.onClusterElementRemoved,
 	})
+	if err != nil {
+		return fmt.Errorf("creation of event handler for %s failed: %w", instance.typeDescription, err)
+	}
 
 	go instance.Run(stop)
 
@@ -69,14 +74,13 @@ func (instance *Definition) HasSynced() bool {
 }
 
 func (instance *Definition) Run(stop support.Channel) {
-	l := instance.log()
 	defer runtime.HandleCrash()
-	l.Info("definition store started")
+	instance.Logger.Info("definition store started")
 
 	go instance.informer.Run(support.ToChan(stop))
 	stop.Wait()
 
-	l.Info("definition store stopped")
+	instance.Logger.Info("definition store stopped")
 }
 
 func (instance *Definition) onClusterElementAdded(new interface{}) {
@@ -85,7 +89,7 @@ func (instance *Definition) onClusterElementAdded(new interface{}) {
 	key, err := cache.MetaNamespaceKeyFunc(new)
 	if err != nil {
 		l.WithError(err).
-			WithField("objectType", reflect.TypeOf(new).String()).
+			With("objectType", reflect.TypeOf(new).String()).
 			Error("cannot determine key of an object of type")
 	}
 
@@ -93,7 +97,7 @@ func (instance *Definition) onClusterElementAdded(new interface{}) {
 		return
 	}
 
-	l = l.WithField("key", key)
+	l = l.With("key", key)
 	if err := instance.OnElementAdded(key, new.(metav1.Object)); err != nil {
 		instance.lastError.Store(err)
 		if instance.OnError != nil {
@@ -113,7 +117,7 @@ func (instance *Definition) onClusterElementUpdated(old interface{}, new interfa
 	key, err := cache.MetaNamespaceKeyFunc(new)
 	if err != nil {
 		l.WithError(err).
-			WithField("objectType", reflect.TypeOf(new).String()).
+			With("objectType", reflect.TypeOf(new).String()).
 			Error("cannot determine key of an object of type")
 	}
 
@@ -121,7 +125,7 @@ func (instance *Definition) onClusterElementUpdated(old interface{}, new interfa
 		return
 	}
 
-	l = l.WithField("key", key)
+	l = l.With("key", key)
 	if err := instance.OnElementUpdated(key, old.(metav1.Object), new.(metav1.Object)); err != nil {
 		instance.lastError.Store(err)
 		if instance.OnError != nil {
@@ -141,7 +145,7 @@ func (instance *Definition) onClusterElementRemoved(old interface{}) {
 	key, err := cache.DeletionHandlingMetaNamespaceKeyFunc(old)
 	if err != nil {
 		l.WithError(err).
-			WithField("objectType", reflect.TypeOf(old).String()).
+			With("objectType", reflect.TypeOf(old).String()).
 			Error("cannot determine key of an object of type")
 	}
 
@@ -149,7 +153,7 @@ func (instance *Definition) onClusterElementRemoved(old interface{}) {
 		return
 	}
 
-	l = l.WithField("key", key)
+	l = l.With("key", key)
 	if err := instance.OnElementRemoved(key); err != nil {
 		instance.lastError.Store(err)
 		if instance.OnError != nil {
@@ -163,14 +167,10 @@ func (instance *Definition) onClusterElementRemoved(old interface{}) {
 	}
 }
 
-func (instance *Definition) log() *log.Entry {
-	return log.WithField("type", instance.typeDescription)
+func (instance *Definition) logEvent(event string) log.Logger {
+	return instance.Logger.With("event", event)
 }
 
-func (instance *Definition) logEvent(event string) *log.Entry {
-	return instance.log().WithField("event", event)
-}
-
-func (instance *Definition) logKey(event string, key string) *log.Entry {
-	return instance.logEvent(event).WithField("key", key)
+func (instance *Definition) logKey(event string, key string) log.Logger {
+	return instance.logEvent(event).With("key", key)
 }
