@@ -3,6 +3,7 @@ package definition
 import (
 	"fmt"
 	"github.com/echocat/lingress/support"
+	log "github.com/echocat/slf4g"
 	"k8s.io/client-go/kubernetes"
 	"time"
 )
@@ -13,12 +14,12 @@ type Definitions struct {
 	Service        *Service
 }
 
-func New(client kubernetes.Interface, resyncAfter time.Duration) (*Definitions, error) {
-	if serviceSecrets, err := NewServiceSecrets(client, resyncAfter); err != nil {
+func New(client kubernetes.Interface, resyncAfter time.Duration, logger log.Logger) (*Definitions, error) {
+	if serviceSecrets, err := NewServiceSecrets(client, resyncAfter, logger); err != nil {
 		return nil, fmt.Errorf("cannot create service secrets definition store: %v", err)
-	} else if ingress, err := NewIngress(client, resyncAfter); err != nil {
+	} else if ingress, err := NewIngress(client, resyncAfter, logger); err != nil {
 		return nil, fmt.Errorf("cannot create ingress definition store: %v", err)
-	} else if service, err := NewService(client, resyncAfter); err != nil {
+	} else if service, err := NewService(client, resyncAfter, logger); err != nil {
 		return nil, fmt.Errorf("cannot create service definition store: %v", err)
 	} else {
 		return &Definitions{
