@@ -2,7 +2,8 @@ package definition
 
 import (
 	"fmt"
-	"k8s.io/api/extensions/v1beta1"
+	log "github.com/echocat/slf4g"
+	networkingv1 "k8s.io/api/networking/v1"
 	"k8s.io/client-go/informers"
 	"k8s.io/client-go/kubernetes"
 	"time"
@@ -12,10 +13,10 @@ type Ingress struct {
 	*Definition
 }
 
-func NewIngress(client kubernetes.Interface, resyncAfter time.Duration) (*Ingress, error) {
+func NewIngress(client kubernetes.Interface, resyncAfter time.Duration, logger log.Logger) (*Ingress, error) {
 	informerFactory := informers.NewSharedInformerFactory(client, resyncAfter)
-	informer := informerFactory.Extensions().V1beta1().Ingresses().Informer()
-	if definition, err := newDefinition("ingress", informer); err != nil {
+	informer := informerFactory.Networking().V1().Ingresses().Informer()
+	if definition, err := newDefinition("ingress", informer, logger); err != nil {
 		return nil, err
 	} else {
 		return &Ingress{
@@ -24,12 +25,12 @@ func NewIngress(client kubernetes.Interface, resyncAfter time.Duration) (*Ingres
 	}
 }
 
-func (instance *Ingress) Get(key string) (*v1beta1.Ingress, error) {
+func (instance *Ingress) Get(key string) (*networkingv1.Ingress, error) {
 	if item, exists, err := instance.informer.GetStore().GetByKey(key); err != nil {
 		return nil, fmt.Errorf("cannot get ingress %s from cache: %v", key, err)
 	} else if !exists {
 		return nil, nil
 	} else {
-		return item.(*v1beta1.Ingress), nil
+		return item.(*networkingv1.Ingress), nil
 	}
 }
