@@ -86,7 +86,7 @@ func (instance *Fallback) RedirectHtml(ctx *context.Context, statusCode int, tar
 		if err := tmpl.Execute(ctx.Client.Response, object); err != nil {
 			ctx.Log().
 				WithError(err).
-				WithField("statusCode", statusCode).
+				With("statusCode", statusCode).
 				Error("could not render status page.")
 		}
 	}
