@@ -1,12 +1,13 @@
 package main
 
 import (
+	"embed"
+	"fmt"
+	"github.com/alecthomas/kingpin/v2"
 	"github.com/echocat/lingress"
 	"github.com/echocat/lingress/support"
-	_ "github.com/echocat/lingress/support"
-	"github.com/gobuffalo/packr"
-	log "github.com/sirupsen/logrus"
-	"gopkg.in/alecthomas/kingpin.v2"
+	_ "github.com/echocat/lingress/support/slf4g_native"
+	"github.com/echocat/slf4g"
 	"os"
 	"os/signal"
 )
@@ -16,14 +17,15 @@ const (
 )
 
 var (
-	localizations = packr.NewBox("../resources/localization")
-	static        = packr.NewBox("../resources/static")
-	templates     = packr.NewBox("../resources/templates")
+	//go:embed localization
+	localizations embed.FS
+	//go:embed templates
+	templates embed.FS
 
 	fps = support.DefaultFileProviders{
-		Localization: localizations,
-		Static:       static,
-		Templates:    templates,
+		Localization: support.FileProviderStrippingPrefix(localizations, "localization"),
+		Static:       support.NoopFileProvider(),
+		Templates:    support.FileProviderStrippingPrefix(templates, "templates"),
 	}
 )
 
@@ -36,6 +38,14 @@ func main() {
 	l, err := lingress.New(fps)
 	support.Must(err)
 
+	app.Flag("version", "Shows the current version information").
+		PreAction(func(*kingpin.ParseContext) error {
+			fmt.Println(rt.LongString())
+			os.Exit(0)
+			return nil
+		}).
+		Bool()
+
 	support.Must(l.RegisterFlag(app, appPrefix))
 	support.MustRegisterGlobalFalgs(app, appPrefix)
 
@@ -45,9 +55,9 @@ func main() {
 		support.ChannelDoOnEvent(stop, func() {
 			close(intSig)
 		})
-		log.WithField("revision", rt.Revision).
-			WithField("branch", rt.Branch).
-			WithField("build", rt.Build).
+		log.With("revision", rt.Revision).
+			With("branch", rt.Branch).
+			With("build", rt.Build).
 			Infof("starting %s...", rt.Name())
 		if err := l.Init(stop); err != nil {
 			return err
