@@ -8,7 +8,7 @@ import (
 func Test_newNode_works_as_expected(t *testing.T) {
 	g := NewGomegaWithT(t)
 
-	actual := newNode()
+	actual := newNode[testValue]()
 
 	g.Expect(actual).ToNot(BeNil())
 	g.Expect(actual.elements).To(BeNil())
