@@ -9,8 +9,8 @@ import (
 func Test_Node_Put_works_with_element_in_root(t *testing.T) {
 	executeTestPutRun(t, []putTestCaseGiven{
 		{path: "a", element: "elementA"},
-	}, node{
-		elements: map[string][]interface{}{
+	}, node[testValue]{
+		elements: map[string][]testValue{
 			"a": {"elementA"},
 		},
 	}, nil)
@@ -19,10 +19,10 @@ func Test_Node_Put_works_with_element_in_root(t *testing.T) {
 func Test_Node_Put_works_with_element_in_subPath(t *testing.T) {
 	executeTestPutRun(t, []putTestCaseGiven{
 		{path: "a/b", element: "elementA"},
-	}, node{
-		children: map[string]*node{
+	}, node[testValue]{
+		children: map[string]*node[testValue]{
 			"a": {
-				elements: map[string][]interface{}{
+				elements: map[string][]testValue{
 					"b": {"elementA"},
 				},
 			},
@@ -38,13 +38,13 @@ func Test_Node_Put_works_with_3_elements_in_root(t *testing.T) {
 		{path: "a", element: "elementA"},
 		{path: "b", element: "elementB"},
 		{path: "c", element: "elementC"},
-	}, node{
-		elements: map[string][]interface{}{
+	}, node[testValue]{
+		elements: map[string][]testValue{
 			"a": {"elementA"},
 			"b": {"elementB"},
 			"c": {"elementC"},
 		},
-	}, &[]interface{}{"elementRoot1", "elementRoot2", "elementRoot3"})
+	}, &[]testValue{"elementRoot1", "elementRoot2", "elementRoot3"})
 }
 
 func Test_Node_Put_works_with_3_elements_across_path(t *testing.T) {
@@ -52,22 +52,22 @@ func Test_Node_Put_works_with_3_elements_across_path(t *testing.T) {
 		{path: "a", element: "elementA"},
 		{path: "a/b", element: "elementAB"},
 		{path: "a/b/c", element: "elementABC"},
-	}, node{
-		children: map[string]*node{
+	}, node[testValue]{
+		children: map[string]*node[testValue]{
 			"a": {
-				children: map[string]*node{
+				children: map[string]*node[testValue]{
 					"b": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c": {"elementABC"},
 						},
 					},
 				},
-				elements: map[string][]interface{}{
+				elements: map[string][]testValue{
 					"b": {"elementAB"},
 				},
 			},
 		},
-		elements: map[string][]interface{}{
+		elements: map[string][]testValue{
 			"a": {"elementA"},
 		},
 	}, nil)
@@ -77,19 +77,19 @@ func Test_Node_Put_works_with_2_elements_across_path(t *testing.T) {
 	executeTestPutRun(t, []putTestCaseGiven{
 		{path: "a", element: "elementA"},
 		{path: "a/b/c", element: "elementABC"},
-	}, node{
-		children: map[string]*node{
+	}, node[testValue]{
+		children: map[string]*node[testValue]{
 			"a": {
-				children: map[string]*node{
+				children: map[string]*node[testValue]{
 					"b": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c": {"elementABC"},
 						},
 					},
 				},
 			},
 		},
-		elements: map[string][]interface{}{
+		elements: map[string][]testValue{
 			"a": {"elementA"},
 		},
 	}, nil)
@@ -101,34 +101,34 @@ func Test_Node_Put_works_with_same_elements_across_path(t *testing.T) {
 		{path: "a/b/c", element: "elementABC!"},
 		{path: "a", element: "elementA"},
 		{path: "a/b/c", element: "elementABC"},
-	}, node{
-		children: map[string]*node{
+	}, node[testValue]{
+		children: map[string]*node[testValue]{
 			"a": {
-				children: map[string]*node{
+				children: map[string]*node[testValue]{
 					"b": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c": {"elementABC!", "elementABC"},
 						},
 					},
 				},
 			},
 		},
-		elements: map[string][]interface{}{
+		elements: map[string][]testValue{
 			"a": {"elementA!", "elementA"},
 		},
 	}, nil)
 }
 
-func executeTestPutRun(t *testing.T, given []putTestCaseGiven, expected node, expectedRootElements *[]interface{}) {
+func executeTestPutRun(t *testing.T, given []putTestCaseGiven, expected node[testValue], expectedRootElements *[]testValue) {
 	g := NewGomegaWithT(t)
 
-	actual := New()
+	actual := New[testValue]()
 	for _, gElement := range given {
 		pathElements := strings.Split(gElement.path, "/")
 		if gElement.path == "" {
 			pathElements = []string{}
 		}
-		err := actual.Put(pathElements, gElement.element)
+		err := actual.Put(pathElements, testValue(gElement.element))
 		g.Expect(err).To(BeNil())
 	}
 
@@ -138,5 +138,5 @@ func executeTestPutRun(t *testing.T, given []putTestCaseGiven, expected node, ex
 
 type putTestCaseGiven struct {
 	path    string
-	element interface{}
+	element string
 }
