@@ -8,555 +8,555 @@ import (
 )
 
 var (
-	givenTreeForRemove = Tree{
-		root: &node{
-			children: map[string]*node{
+	givenTreeForRemove = Tree[testValue]{
+		root: &node[testValue]{
+			children: map[string]*node[testValue]{
 				"a1": {
-					children: map[string]*node{
+					children: map[string]*node[testValue]{
 						"b1": {
-							elements: map[string][]interface{}{
+							elements: map[string][]testValue{
 								"c1": {"elementA1B1C1!", "elementA1B1C1?", "elementA1B1C1-"},
 								"c2": {"elementA1B1C2!", "elementA1B1C2?", "elementA1B1C2-"},
 								"c3": {"elementA1B1C3!", "elementA1B1C3?", "elementA1B1C3-"},
 							},
 						},
 						"b2": {
-							elements: map[string][]interface{}{
+							elements: map[string][]testValue{
 								"c1": {"elementA1B2C1!", "elementA1B2C1?", "elementA1B2C1-"},
 								"c2": {"elementA1B2C2!", "elementA1B2C2?", "elementA1B2C2-"},
 								"c3": {"elementA1B2C3!", "elementA1B2C3?", "elementA1B2C3-"},
 							},
 						},
 					},
-					elements: map[string][]interface{}{
+					elements: map[string][]testValue{
 						"b1": {"elementA1B1!", "elementA1B1?", "elementA1B2-"},
 						"b2": {"elementA1B2!", "elementA1B2?", "elementA1B2-"},
 						"b3": {"elementA1B3!", "elementA1B3?", "elementA1B3-"},
 					},
 				},
 				"a2": {
-					children: map[string]*node{
+					children: map[string]*node[testValue]{
 						"b1": {
-							elements: map[string][]interface{}{
+							elements: map[string][]testValue{
 								"c1": {"elementA2B1C1!", "elementA2B1C1?", "elementA2B1C1-"},
 								"c2": {"elementA2B1C2!", "elementA2B1C2?", "elementA2B1C2-"},
 								"c3": {"elementA2B1C3!", "elementA2B1C3?", "elementA2B1C3-"},
 							},
 						},
 						"b2": {
-							elements: map[string][]interface{}{
+							elements: map[string][]testValue{
 								"c1": {"elementA2B2C1!", "elementA2B2C1?", "elementA2B2C1-"},
 								"c2": {"elementA2B2C2!", "elementA2B2C2?", "elementA2B2C2-"},
 								"c3": {"elementA2B2C3!", "elementA2B2C3?", "elementA2B2C3-"},
 							},
 						},
 					},
-					elements: map[string][]interface{}{
+					elements: map[string][]testValue{
 						"b1": {"elementA2B1!", "elementA2B1?", "elementA2B2-"},
 						"b2": {"elementA2B2!", "elementA2B2?", "elementA2B2-"},
 						"b3": {"elementA2B3!", "elementA2B3?", "elementA2B3-"},
 					},
 				},
 			},
-			elements: map[string][]interface{}{
+			elements: map[string][]testValue{
 				"a1": {"elementA1!", "elementA1?", "elementA1-"},
 				"a2": {"elementA2!", "elementA2?", "elementA2-"},
 				"a3": {"elementA3!", "elementA3?", "elementA3-"},
 			},
 		},
-		rootElements: &[]interface{}{"element!", "element?", "element-"},
+		rootElements: &[]testValue{"element!", "element?", "element-"},
 	}
 )
 
 func Test_Node_Remove_removes_elements_from_start_of_groups_across_tree(t *testing.T) {
-	executeTestRemoveRun(t, func(path []string, element interface{}) bool {
-		return strings.HasSuffix(element.(string), "!")
-	}, node{
-		children: map[string]*node{
+	executeTestRemoveRun(t, func(path []string, element testValue) bool {
+		return strings.HasSuffix(string(element), "!")
+	}, node[testValue]{
+		children: map[string]*node[testValue]{
 			"a1": {
-				children: map[string]*node{
+				children: map[string]*node[testValue]{
 					"b1": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA1B1C1?", "elementA1B1C1-"},
 							"c2": {"elementA1B1C2?", "elementA1B1C2-"},
 							"c3": {"elementA1B1C3?", "elementA1B1C3-"},
 						},
 					},
 					"b2": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA1B2C1?", "elementA1B2C1-"},
 							"c2": {"elementA1B2C2?", "elementA1B2C2-"},
 							"c3": {"elementA1B2C3?", "elementA1B2C3-"},
 						},
 					},
 				},
-				elements: map[string][]interface{}{
+				elements: map[string][]testValue{
 					"b1": {"elementA1B1?", "elementA1B2-"},
 					"b2": {"elementA1B2?", "elementA1B2-"},
 					"b3": {"elementA1B3?", "elementA1B3-"},
 				},
 			},
 			"a2": {
-				children: map[string]*node{
+				children: map[string]*node[testValue]{
 					"b1": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA2B1C1?", "elementA2B1C1-"},
 							"c2": {"elementA2B1C2?", "elementA2B1C2-"},
 							"c3": {"elementA2B1C3?", "elementA2B1C3-"},
 						},
 					},
 					"b2": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA2B2C1?", "elementA2B2C1-"},
 							"c2": {"elementA2B2C2?", "elementA2B2C2-"},
 							"c3": {"elementA2B2C3?", "elementA2B2C3-"},
 						},
 					},
 				},
-				elements: map[string][]interface{}{
+				elements: map[string][]testValue{
 					"b1": {"elementA2B1?", "elementA2B2-"},
 					"b2": {"elementA2B2?", "elementA2B2-"},
 					"b3": {"elementA2B3?", "elementA2B3-"},
 				},
 			},
 		},
-		elements: map[string][]interface{}{
+		elements: map[string][]testValue{
 			"a1": {"elementA1?", "elementA1-"},
 			"a2": {"elementA2?", "elementA2-"},
 			"a3": {"elementA3?", "elementA3-"},
 		},
-	}, &[]interface{}{"element?", "element-"})
+	}, &[]testValue{"element?", "element-"})
 }
 
 func Test_Node_Remove_removes_elements_from_end_of_groups_across_tree(t *testing.T) {
-	executeTestRemoveRun(t, func(path []string, element interface{}) bool {
-		return strings.HasSuffix(element.(string), "-")
-	}, node{
-		children: map[string]*node{
+	executeTestRemoveRun(t, func(path []string, element testValue) bool {
+		return strings.HasSuffix(string(element), "-")
+	}, node[testValue]{
+		children: map[string]*node[testValue]{
 			"a1": {
-				children: map[string]*node{
+				children: map[string]*node[testValue]{
 					"b1": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA1B1C1!", "elementA1B1C1?"},
 							"c2": {"elementA1B1C2!", "elementA1B1C2?"},
 							"c3": {"elementA1B1C3!", "elementA1B1C3?"},
 						},
 					},
 					"b2": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA1B2C1!", "elementA1B2C1?"},
 							"c2": {"elementA1B2C2!", "elementA1B2C2?"},
 							"c3": {"elementA1B2C3!", "elementA1B2C3?"},
 						},
 					},
 				},
-				elements: map[string][]interface{}{
+				elements: map[string][]testValue{
 					"b1": {"elementA1B1!", "elementA1B1?"},
 					"b2": {"elementA1B2!", "elementA1B2?"},
 					"b3": {"elementA1B3!", "elementA1B3?"},
 				},
 			},
 			"a2": {
-				children: map[string]*node{
+				children: map[string]*node[testValue]{
 					"b1": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA2B1C1!", "elementA2B1C1?"},
 							"c2": {"elementA2B1C2!", "elementA2B1C2?"},
 							"c3": {"elementA2B1C3!", "elementA2B1C3?"},
 						},
 					},
 					"b2": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA2B2C1!", "elementA2B2C1?"},
 							"c2": {"elementA2B2C2!", "elementA2B2C2?"},
 							"c3": {"elementA2B2C3!", "elementA2B2C3?"},
 						},
 					},
 				},
-				elements: map[string][]interface{}{
+				elements: map[string][]testValue{
 					"b1": {"elementA2B1!", "elementA2B1?"},
 					"b2": {"elementA2B2!", "elementA2B2?"},
 					"b3": {"elementA2B3!", "elementA2B3?"},
 				},
 			},
 		},
-		elements: map[string][]interface{}{
+		elements: map[string][]testValue{
 			"a1": {"elementA1!", "elementA1?"},
 			"a2": {"elementA2!", "elementA2?"},
 			"a3": {"elementA3!", "elementA3?"},
 		},
-	}, &[]interface{}{"element!", "element?"})
+	}, &[]testValue{"element!", "element?"})
 }
 
 func Test_Node_Remove_removes_elements_from_center_of_groups_across_tree(t *testing.T) {
-	executeTestRemoveRun(t, func(path []string, element interface{}) bool {
-		return strings.HasSuffix(element.(string), "?")
-	}, node{
-		children: map[string]*node{
+	executeTestRemoveRun(t, func(path []string, element testValue) bool {
+		return strings.HasSuffix(string(element), "?")
+	}, node[testValue]{
+		children: map[string]*node[testValue]{
 			"a1": {
-				children: map[string]*node{
+				children: map[string]*node[testValue]{
 					"b1": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA1B1C1!", "elementA1B1C1-"},
 							"c2": {"elementA1B1C2!", "elementA1B1C2-"},
 							"c3": {"elementA1B1C3!", "elementA1B1C3-"},
 						},
 					},
 					"b2": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA1B2C1!", "elementA1B2C1-"},
 							"c2": {"elementA1B2C2!", "elementA1B2C2-"},
 							"c3": {"elementA1B2C3!", "elementA1B2C3-"},
 						},
 					},
 				},
-				elements: map[string][]interface{}{
+				elements: map[string][]testValue{
 					"b1": {"elementA1B1!", "elementA1B2-"},
 					"b2": {"elementA1B2!", "elementA1B2-"},
 					"b3": {"elementA1B3!", "elementA1B3-"},
 				},
 			},
 			"a2": {
-				children: map[string]*node{
+				children: map[string]*node[testValue]{
 					"b1": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA2B1C1!", "elementA2B1C1-"},
 							"c2": {"elementA2B1C2!", "elementA2B1C2-"},
 							"c3": {"elementA2B1C3!", "elementA2B1C3-"},
 						},
 					},
 					"b2": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA2B2C1!", "elementA2B2C1-"},
 							"c2": {"elementA2B2C2!", "elementA2B2C2-"},
 							"c3": {"elementA2B2C3!", "elementA2B2C3-"},
 						},
 					},
 				},
-				elements: map[string][]interface{}{
+				elements: map[string][]testValue{
 					"b1": {"elementA2B1!", "elementA2B2-"},
 					"b2": {"elementA2B2!", "elementA2B2-"},
 					"b3": {"elementA2B3!", "elementA2B3-"},
 				},
 			},
 		},
-		elements: map[string][]interface{}{
+		elements: map[string][]testValue{
 			"a1": {"elementA1!", "elementA1-"},
 			"a2": {"elementA2!", "elementA2-"},
 			"a3": {"elementA3!", "elementA3-"},
 		},
-	}, &[]interface{}{"element!", "element-"})
+	}, &[]testValue{"element!", "element-"})
 }
 
 func Test_Node_Remove_removes_elements_from_beginning_of_tree(t *testing.T) {
-	executeTestRemoveRun(t, func(path []string, element interface{}) bool {
+	executeTestRemoveRun(t, func(path []string, element testValue) bool {
 		return reflect.DeepEqual(path, []string{"a1"}) || reflect.DeepEqual(path, []string{"a1", "b1"}) || reflect.DeepEqual(path, []string{"a1", "b1", "c1"})
-	}, node{
-		children: map[string]*node{
+	}, node[testValue]{
+		children: map[string]*node[testValue]{
 			"a1": {
-				children: map[string]*node{
+				children: map[string]*node[testValue]{
 					"b1": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c2": {"elementA1B1C2!", "elementA1B1C2?", "elementA1B1C2-"},
 							"c3": {"elementA1B1C3!", "elementA1B1C3?", "elementA1B1C3-"},
 						},
 					},
 					"b2": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA1B2C1!", "elementA1B2C1?", "elementA1B2C1-"},
 							"c2": {"elementA1B2C2!", "elementA1B2C2?", "elementA1B2C2-"},
 							"c3": {"elementA1B2C3!", "elementA1B2C3?", "elementA1B2C3-"},
 						},
 					},
 				},
-				elements: map[string][]interface{}{
+				elements: map[string][]testValue{
 					"b2": {"elementA1B2!", "elementA1B2?", "elementA1B2-"},
 					"b3": {"elementA1B3!", "elementA1B3?", "elementA1B3-"},
 				},
 			},
 			"a2": {
-				children: map[string]*node{
+				children: map[string]*node[testValue]{
 					"b1": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA2B1C1!", "elementA2B1C1?", "elementA2B1C1-"},
 							"c2": {"elementA2B1C2!", "elementA2B1C2?", "elementA2B1C2-"},
 							"c3": {"elementA2B1C3!", "elementA2B1C3?", "elementA2B1C3-"},
 						},
 					},
 					"b2": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA2B2C1!", "elementA2B2C1?", "elementA2B2C1-"},
 							"c2": {"elementA2B2C2!", "elementA2B2C2?", "elementA2B2C2-"},
 							"c3": {"elementA2B2C3!", "elementA2B2C3?", "elementA2B2C3-"},
 						},
 					},
 				},
-				elements: map[string][]interface{}{
+				elements: map[string][]testValue{
 					"b1": {"elementA2B1!", "elementA2B1?", "elementA2B2-"},
 					"b2": {"elementA2B2!", "elementA2B2?", "elementA2B2-"},
 					"b3": {"elementA2B3!", "elementA2B3?", "elementA2B3-"},
 				},
 			},
 		},
-		elements: map[string][]interface{}{
+		elements: map[string][]testValue{
 			"a2": {"elementA2!", "elementA2?", "elementA2-"},
 			"a3": {"elementA3!", "elementA3?", "elementA3-"},
 		},
-	}, &[]interface{}{"element!", "element?", "element-"})
+	}, &[]testValue{"element!", "element?", "element-"})
 }
 
 func Test_Node_Remove_removes_elements_from_end_of_tree(t *testing.T) {
-	executeTestRemoveRun(t, func(path []string, element interface{}) bool {
+	executeTestRemoveRun(t, func(path []string, element testValue) bool {
 		return reflect.DeepEqual(path, []string{"a3"}) || reflect.DeepEqual(path, []string{"a1", "b3"}) || reflect.DeepEqual(path, []string{"a1", "b1", "c3"})
-	}, node{
-		children: map[string]*node{
+	}, node[testValue]{
+		children: map[string]*node[testValue]{
 			"a1": {
-				children: map[string]*node{
+				children: map[string]*node[testValue]{
 					"b1": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA1B1C1!", "elementA1B1C1?", "elementA1B1C1-"},
 							"c2": {"elementA1B1C2!", "elementA1B1C2?", "elementA1B1C2-"},
 						},
 					},
 					"b2": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA1B2C1!", "elementA1B2C1?", "elementA1B2C1-"},
 							"c2": {"elementA1B2C2!", "elementA1B2C2?", "elementA1B2C2-"},
 							"c3": {"elementA1B2C3!", "elementA1B2C3?", "elementA1B2C3-"},
 						},
 					},
 				},
-				elements: map[string][]interface{}{
+				elements: map[string][]testValue{
 					"b1": {"elementA1B1!", "elementA1B1?", "elementA1B2-"},
 					"b2": {"elementA1B2!", "elementA1B2?", "elementA1B2-"},
 				},
 			},
 			"a2": {
-				children: map[string]*node{
+				children: map[string]*node[testValue]{
 					"b1": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA2B1C1!", "elementA2B1C1?", "elementA2B1C1-"},
 							"c2": {"elementA2B1C2!", "elementA2B1C2?", "elementA2B1C2-"},
 							"c3": {"elementA2B1C3!", "elementA2B1C3?", "elementA2B1C3-"},
 						},
 					},
 					"b2": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA2B2C1!", "elementA2B2C1?", "elementA2B2C1-"},
 							"c2": {"elementA2B2C2!", "elementA2B2C2?", "elementA2B2C2-"},
 							"c3": {"elementA2B2C3!", "elementA2B2C3?", "elementA2B2C3-"},
 						},
 					},
 				},
-				elements: map[string][]interface{}{
+				elements: map[string][]testValue{
 					"b1": {"elementA2B1!", "elementA2B1?", "elementA2B2-"},
 					"b2": {"elementA2B2!", "elementA2B2?", "elementA2B2-"},
 					"b3": {"elementA2B3!", "elementA2B3?", "elementA2B3-"},
 				},
 			},
 		},
-		elements: map[string][]interface{}{
+		elements: map[string][]testValue{
 			"a1": {"elementA1!", "elementA1?", "elementA1-"},
 			"a2": {"elementA2!", "elementA2?", "elementA2-"},
 		},
-	}, &[]interface{}{"element!", "element?", "element-"})
+	}, &[]testValue{"element!", "element?", "element-"})
 }
 
 func Test_Node_Remove_removes_elements_from_center_of_tree(t *testing.T) {
-	executeTestRemoveRun(t, func(path []string, element interface{}) bool {
+	executeTestRemoveRun(t, func(path []string, element testValue) bool {
 		return reflect.DeepEqual(path, []string{"a2"}) || reflect.DeepEqual(path, []string{"a1", "b2"}) || reflect.DeepEqual(path, []string{"a1", "b1", "c2"})
-	}, node{
-		children: map[string]*node{
+	}, node[testValue]{
+		children: map[string]*node[testValue]{
 			"a1": {
-				children: map[string]*node{
+				children: map[string]*node[testValue]{
 					"b1": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA1B1C1!", "elementA1B1C1?", "elementA1B1C1-"},
 							"c3": {"elementA1B1C3!", "elementA1B1C3?", "elementA1B1C3-"},
 						},
 					},
 					"b2": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA1B2C1!", "elementA1B2C1?", "elementA1B2C1-"},
 							"c2": {"elementA1B2C2!", "elementA1B2C2?", "elementA1B2C2-"},
 							"c3": {"elementA1B2C3!", "elementA1B2C3?", "elementA1B2C3-"},
 						},
 					},
 				},
-				elements: map[string][]interface{}{
+				elements: map[string][]testValue{
 					"b1": {"elementA1B1!", "elementA1B1?", "elementA1B2-"},
 					"b3": {"elementA1B3!", "elementA1B3?", "elementA1B3-"},
 				},
 			},
 			"a2": {
-				children: map[string]*node{
+				children: map[string]*node[testValue]{
 					"b1": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA2B1C1!", "elementA2B1C1?", "elementA2B1C1-"},
 							"c2": {"elementA2B1C2!", "elementA2B1C2?", "elementA2B1C2-"},
 							"c3": {"elementA2B1C3!", "elementA2B1C3?", "elementA2B1C3-"},
 						},
 					},
 					"b2": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA2B2C1!", "elementA2B2C1?", "elementA2B2C1-"},
 							"c2": {"elementA2B2C2!", "elementA2B2C2?", "elementA2B2C2-"},
 							"c3": {"elementA2B2C3!", "elementA2B2C3?", "elementA2B2C3-"},
 						},
 					},
 				},
-				elements: map[string][]interface{}{
+				elements: map[string][]testValue{
 					"b1": {"elementA2B1!", "elementA2B1?", "elementA2B2-"},
 					"b2": {"elementA2B2!", "elementA2B2?", "elementA2B2-"},
 					"b3": {"elementA2B3!", "elementA2B3?", "elementA2B3-"},
 				},
 			},
 		},
-		elements: map[string][]interface{}{
+		elements: map[string][]testValue{
 			"a1": {"elementA1!", "elementA1?", "elementA1-"},
 			"a3": {"elementA3!", "elementA3?", "elementA3-"},
 		},
-	}, &[]interface{}{"element!", "element?", "element-"})
+	}, &[]testValue{"element!", "element?", "element-"})
 }
 
 func Test_Node_Remove_removes_whole_branch_at_end_of_tree(t *testing.T) {
-	executeTestRemoveRun(t, func(path []string, element interface{}) bool {
+	executeTestRemoveRun(t, func(path []string, element testValue) bool {
 		return len(path) == 3 &&
 			(path[0] == "a1" || path[0] == "a2") &&
 			path[1] == "b1"
-	}, node{
-		children: map[string]*node{
+	}, node[testValue]{
+		children: map[string]*node[testValue]{
 			"a1": {
-				children: map[string]*node{
+				children: map[string]*node[testValue]{
 					"b2": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA1B2C1!", "elementA1B2C1?", "elementA1B2C1-"},
 							"c2": {"elementA1B2C2!", "elementA1B2C2?", "elementA1B2C2-"},
 							"c3": {"elementA1B2C3!", "elementA1B2C3?", "elementA1B2C3-"},
 						},
 					},
 				},
-				elements: map[string][]interface{}{
+				elements: map[string][]testValue{
 					"b1": {"elementA1B1!", "elementA1B1?", "elementA1B2-"},
 					"b2": {"elementA1B2!", "elementA1B2?", "elementA1B2-"},
 					"b3": {"elementA1B3!", "elementA1B3?", "elementA1B3-"},
 				},
 			},
 			"a2": {
-				children: map[string]*node{
+				children: map[string]*node[testValue]{
 					"b2": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA2B2C1!", "elementA2B2C1?", "elementA2B2C1-"},
 							"c2": {"elementA2B2C2!", "elementA2B2C2?", "elementA2B2C2-"},
 							"c3": {"elementA2B2C3!", "elementA2B2C3?", "elementA2B2C3-"},
 						},
 					},
 				},
-				elements: map[string][]interface{}{
+				elements: map[string][]testValue{
 					"b1": {"elementA2B1!", "elementA2B1?", "elementA2B2-"},
 					"b2": {"elementA2B2!", "elementA2B2?", "elementA2B2-"},
 					"b3": {"elementA2B3!", "elementA2B3?", "elementA2B3-"},
 				},
 			},
 		},
-		elements: map[string][]interface{}{
+		elements: map[string][]testValue{
 			"a1": {"elementA1!", "elementA1?", "elementA1-"},
 			"a2": {"elementA2!", "elementA2?", "elementA2-"},
 			"a3": {"elementA3!", "elementA3?", "elementA3-"},
 		},
-	}, &[]interface{}{"element!", "element?", "element-"})
+	}, &[]testValue{"element!", "element?", "element-"})
 }
 
 func Test_Node_Remove_removes_whole_branch_at_center_of_tree(t *testing.T) {
-	executeTestRemoveRun(t, func(path []string, element interface{}) bool {
+	executeTestRemoveRun(t, func(path []string, element testValue) bool {
 		return len(path) >= 1 &&
 			path[0] == "a1"
-	}, node{
-		children: map[string]*node{
+	}, node[testValue]{
+		children: map[string]*node[testValue]{
 			"a2": {
-				children: map[string]*node{
+				children: map[string]*node[testValue]{
 					"b1": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA2B1C1!", "elementA2B1C1?", "elementA2B1C1-"},
 							"c2": {"elementA2B1C2!", "elementA2B1C2?", "elementA2B1C2-"},
 							"c3": {"elementA2B1C3!", "elementA2B1C3?", "elementA2B1C3-"},
 						},
 					},
 					"b2": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA2B2C1!", "elementA2B2C1?", "elementA2B2C1-"},
 							"c2": {"elementA2B2C2!", "elementA2B2C2?", "elementA2B2C2-"},
 							"c3": {"elementA2B2C3!", "elementA2B2C3?", "elementA2B2C3-"},
 						},
 					},
 				},
-				elements: map[string][]interface{}{
+				elements: map[string][]testValue{
 					"b1": {"elementA2B1!", "elementA2B1?", "elementA2B2-"},
 					"b2": {"elementA2B2!", "elementA2B2?", "elementA2B2-"},
 					"b3": {"elementA2B3!", "elementA2B3?", "elementA2B3-"},
 				},
 			},
 		},
-		elements: map[string][]interface{}{
+		elements: map[string][]testValue{
 			"a2": {"elementA2!", "elementA2?", "elementA2-"},
 			"a3": {"elementA3!", "elementA3?", "elementA3-"},
 		},
-	}, &[]interface{}{"element!", "element?", "element-"})
+	}, &[]testValue{"element!", "element?", "element-"})
 }
 
 func Test_Node_Remove_removes_just_rootElements(t *testing.T) {
-	executeTestRemoveRun(t, func(path []string, element interface{}) bool {
+	executeTestRemoveRun(t, func(path []string, element testValue) bool {
 		return len(path) == 0
-	}, node{
-		children: map[string]*node{
+	}, node[testValue]{
+		children: map[string]*node[testValue]{
 			"a1": {
-				children: map[string]*node{
+				children: map[string]*node[testValue]{
 					"b1": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA1B1C1!", "elementA1B1C1?", "elementA1B1C1-"},
 							"c2": {"elementA1B1C2!", "elementA1B1C2?", "elementA1B1C2-"},
 							"c3": {"elementA1B1C3!", "elementA1B1C3?", "elementA1B1C3-"},
 						},
 					},
 					"b2": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA1B2C1!", "elementA1B2C1?", "elementA1B2C1-"},
 							"c2": {"elementA1B2C2!", "elementA1B2C2?", "elementA1B2C2-"},
 							"c3": {"elementA1B2C3!", "elementA1B2C3?", "elementA1B2C3-"},
 						},
 					},
 				},
-				elements: map[string][]interface{}{
+				elements: map[string][]testValue{
 					"b1": {"elementA1B1!", "elementA1B1?", "elementA1B2-"},
 					"b2": {"elementA1B2!", "elementA1B2?", "elementA1B2-"},
 					"b3": {"elementA1B3!", "elementA1B3?", "elementA1B3-"},
 				},
 			},
 			"a2": {
-				children: map[string]*node{
+				children: map[string]*node[testValue]{
 					"b1": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA2B1C1!", "elementA2B1C1?", "elementA2B1C1-"},
 							"c2": {"elementA2B1C2!", "elementA2B1C2?", "elementA2B1C2-"},
 							"c3": {"elementA2B1C3!", "elementA2B1C3?", "elementA2B1C3-"},
 						},
 					},
 					"b2": {
-						elements: map[string][]interface{}{
+						elements: map[string][]testValue{
 							"c1": {"elementA2B2C1!", "elementA2B2C1?", "elementA2B2C1-"},
 							"c2": {"elementA2B2C2!", "elementA2B2C2?", "elementA2B2C2-"},
 							"c3": {"elementA2B2C3!", "elementA2B2C3?", "elementA2B2C3-"},
 						},
 					},
 				},
-				elements: map[string][]interface{}{
+				elements: map[string][]testValue{
 					"b1": {"elementA2B1!", "elementA2B1?", "elementA2B2-"},
 					"b2": {"elementA2B2!", "elementA2B2?", "elementA2B2-"},
 					"b3": {"elementA2B3!", "elementA2B3?", "elementA2B3-"},
 				},
 			},
 		},
-		elements: map[string][]interface{}{
+		elements: map[string][]testValue{
 			"a1": {"elementA1!", "elementA1?", "elementA1-"},
 			"a2": {"elementA2!", "elementA2?", "elementA2-"},
 			"a3": {"elementA3!", "elementA3?", "elementA3-"},
@@ -565,12 +565,12 @@ func Test_Node_Remove_removes_just_rootElements(t *testing.T) {
 }
 
 func Test_Node_Remove_removes_whole_tree(t *testing.T) {
-	executeTestRemoveRun(t, func(path []string, element interface{}) bool {
+	executeTestRemoveRun(t, func(path []string, element testValue) bool {
 		return true
-	}, node{}, nil)
+	}, node[testValue]{}, nil)
 }
 
-func executeTestRemoveRun(t *testing.T, predicate Predicate, expected node, expectedRootElements *[]interface{}) {
+func executeTestRemoveRun(t *testing.T, predicate Predicate[testValue], expected node[testValue], expectedRootElements *[]testValue) {
 	g := NewGomegaWithT(t)
 
 	instance := givenTreeForRemove.Clone()
