@@ -2,12 +2,13 @@ package rules
 
 import (
 	"errors"
+	"fmt"
+	networkingv1 "k8s.io/api/networking/v1"
 	"strings"
 )
 
 const (
-	annotationIngressClass = "kubernetes.io/ingress.class"
-	ingressClass           = "lingress"
+	ingressClass = "lingress"
 )
 
 var (
@@ -23,12 +24,12 @@ func normalizeHostname(in string) string {
 func ParsePath(in string, faultTolerant bool) ([]string, error) {
 	out := strings.Split(in, "/")
 	if len(out) == 1 && !faultTolerant {
-		return nil, ErrIllegalPath
+		return nil, fmt.Errorf("%w: %s", ErrIllegalPath, in)
 	}
 	if len(out) > 0 && out[0] == "" {
 		out = out[1:]
 	} else {
-		return nil, ErrIllegalPath
+		return nil, fmt.Errorf("%w: %s", ErrIllegalPath, in)
 	}
 
 	if len(out) == 1 && out[0] == "" {
@@ -41,8 +42,22 @@ func ParsePath(in string, faultTolerant bool) ([]string, error) {
 
 	for _, element := range out {
 		if element == "" && !faultTolerant {
-			return nil, ErrIllegalPath
+			return nil, fmt.Errorf("%w: %s", ErrIllegalPath, in)
 		}
 	}
 	return out, nil
+}
+
+func ParsePathType(in *networkingv1.PathType) (PathType, error) {
+	if in == nil {
+		return PathTypePrefix, nil
+	}
+	switch *in {
+	case "Prefix", "ImplementationSpecific":
+		return PathTypePrefix, nil
+	case "Exact":
+		return PathTypeExact, nil
+	default:
+		return 0, fmt.Errorf("cannot handle path type: %s", *in)
+	}
 }
