@@ -1,7 +1,7 @@
 package support
 
 import (
-	"gopkg.in/alecthomas/kingpin.v2"
+	"github.com/alecthomas/kingpin/v2"
 	"strings"
 	"unicode"
 )
